Put in Architecture layout & Fixed Up data structures
I wrote in the main component of the algorithm for building the trees
and linked list. I also fixed some of the utility functions and the
linked list implementation.

lmk if you see anything you want to discuss or have questions about.

I'll work on filling in some of the TODOs next.